--- a/src/Xamarin.Android.Build.Tasks/Tasks/GeneratePackageManagerJava.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tasks/GeneratePackageManagerJava.cs
@@ -28,7 +28,6 @@ namespace Xamarin.Android.Tasks
 		[Required]
 		public ITaskItem[] ResolvedUserAssemblies { get; set; }
 
-		[Required]
 		public ITaskItem[] NativeLibraries { get; set; }
 
 		public ITaskItem[] MonoComponents { get; set; }
@@ -355,16 +354,18 @@ namespace Xamarin.Android.Tasks
 			}
 
 			var uniqueNativeLibraries = new List<ITaskItem> ();
-			var seenNativeLibraryNames = new HashSet<string> (StringComparer.OrdinalIgnoreCase);
-			foreach (ITaskItem item in NativeLibraries) {
-				// We don't care about different ABIs here, just the file name
-				string name = Path.GetFileName (item.ItemSpec);
-				if (seenNativeLibraryNames.Contains (name)) {
-					continue;
-				}
+			if (NativeLibraries != null) {
+				var seenNativeLibraryNames = new HashSet<string> (StringComparer.OrdinalIgnoreCase);
+				foreach (ITaskItem item in NativeLibraries) {
+					// We don't care about different ABIs here, just the file name
+					string name = Path.GetFileName (item.ItemSpec);
+					if (seenNativeLibraryNames.Contains (name)) {
+						continue;
+					}
 
-				seenNativeLibraryNames.Add (name);
-				uniqueNativeLibraries.Add (item);
+					seenNativeLibraryNames.Add (name);
+					uniqueNativeLibraries.Add (item);
+				}
 			}
 
 			bool haveRuntimeConfigBlob = !String.IsNullOrEmpty (RuntimeConfigBinFilePath) && File.Exists (RuntimeConfigBinFilePath);