--- a/src/Xamarin.Android.Build.Tasks/Utilities/CompressedAssembliesNativeAssemblyGenerator.cs
+++ b/src/Xamarin.Android.Build.Tasks/Utilities/CompressedAssembliesNativeAssemblyGenerator.cs
@@ -6,7 +6,7 @@ using Xamarin.Android.Tools;
 
 namespace Xamarin.Android.Tasks
 {
-	class CompressedAssembliesNativeAssemblyGenerator
+	class CompressedAssembliesNativeAssemblyGenerator : NativeAssemblyComposer
 	{
 
 		// Order of fields and their type must correspond *exactly* to that in
@@ -16,7 +16,7 @@ namespace Xamarin.Android.Tasks
 			public uint   uncompressed_file_size;
 			public bool   loaded;
 
-			[NativeAssembler (PointerToSymbol = true)]
+			[NativeAssemblerString (PointerToSymbol = true)]
 			public string data = String.Empty;
 		};
 
@@ -26,7 +26,7 @@ namespace Xamarin.Android.Tasks
 		{
 			public uint count;
 
-			[NativeAssembler (PointerToSymbol = true)]
+			[NativeAssemblerString (PointerToSymbol = true)]
 			public string descriptors;
 		};
 
@@ -36,19 +36,15 @@ namespace Xamarin.Android.Tasks
 		IDictionary<string, CompressedAssemblyInfo> assemblies;
 		string dataIncludeFile;
 
-		AndroidTargetArch arch;
-
 		public CompressedAssembliesNativeAssemblyGenerator (AndroidTargetArch arch, IDictionary<string, CompressedAssemblyInfo> assemblies, string baseFilePath)
+			: base (arch)
 		{
 			this.assemblies = assemblies;
-			this.arch = arch;
 			dataIncludeFile = $"{baseFilePath}-data.inc";
 		}
 
-		public void Write (StreamWriter output, string fileName)
+		protected override void Write (NativeAssemblyGenerator generator)
 		{
-			NativeAssemblyGenerator generator = NativeAssemblyGenerator.Create (arch, output, fileName);
-
 			if (assemblies == null || assemblies.Count == 0) {
 				WriteCompressedAssembliesStructure (generator, 0, null);
 				return;
@@ -56,8 +52,8 @@ namespace Xamarin.Android.Tasks
 
 			generator.WriteInclude (Path.GetFileName (dataIncludeFile));
 
-			string label = generator.MakeLocalLabel (DescriptorsField);
-			using (var dataOutput = MemoryStreamPool.Shared.CreateStreamWriter (output.Encoding)) {
+			string label;
+			using (var dataOutput = MemoryStreamPool.Shared.CreateStreamWriter (generator.Output.Encoding)) {
 				generator.WriteDataSection ();
 
 				var descriptor = new CompressedAssemblyDescriptor {
@@ -76,7 +72,7 @@ namespace Xamarin.Android.Tasks
 					NativeAssemblyGenerator.StructureWriteContext descriptorStruct = generator.AddStructureArrayElement (descriptorArray);
 					generator.WriteStructure (descriptorStruct, descriptor);
 				}
-				generator.WriteSymbol (descriptorArray, label);
+				label = generator.WriteSymbol (descriptorArray, DescriptorsField);
 
 				dataOutput.Flush ();
 				Files.CopyIfStreamChanged (dataOutput.BaseStream, dataIncludeFile);