--- a/src/Xamarin.Android.Build.Tasks/Utilities/ELFHelper.cs
+++ b/src/Xamarin.Android.Build.Tasks/Utilities/ELFHelper.cs
@@ -8,6 +8,8 @@ using ELFSharp.ELF.Sections;
 using Microsoft.Android.Build.Tasks;
 using Microsoft.Build.Utilities;
 
+using ELFSymbolType = global::ELFSharp.ELF.Sections.SymbolType;
+
 namespace Xamarin.Android.Tasks
 {
 	static class ELFHelper
@@ -46,7 +48,7 @@ namespace Xamarin.Android.Tasks
 
 			bool mono_aot_file_info_found = false;
 			foreach (var entry in symtab.Entries) {
-				if (String.Compare ("mono_aot_file_info", entry.Name, StringComparison.Ordinal) == 0 && entry.Type == SymbolType.Object) {
+				if (String.Compare ("mono_aot_file_info", entry.Name, StringComparison.Ordinal) == 0 && entry.Type == ELFSymbolType.Object) {
 					mono_aot_file_info_found = true;
 					break;
 				}
@@ -67,7 +69,7 @@ namespace Xamarin.Android.Tasks
 
 			bool isElf64 = elf.Class == Class.Bit64;
 			foreach (var entry in symtab.Entries) {
-				if (entry.Type == SymbolType.Function) {
+				if (entry.Type == ELFSymbolType.Function) {
 					return false;
 				}