… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Utilities/EnvironmentHelper.cs
… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Xamarin.Android.Build.Tests.csproj
… src/Xamarin.Android.Build.Tasks/Utilities/ARMNativeAssemblerTargetProvider.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblerTargetProvider.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/Arm64NativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/ArmNativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/NativeAssemblerAttribute.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/NativeAssemblerContextDataProvider.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/NativeAssemblyComposer.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/NativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/SectionFlags.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/SectionType.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/SymbolType.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/X86NativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/X86_32NativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/X86_64NativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/TypeMapGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/TypeMappingAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/TypeMappingDebugNativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/TypeMappingReleaseNativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/X86NativeAssemblerTargetProvider.cs
… src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets
… src/monodroid/jni/application_dso_stub.cc
… src/monodroid/jni/embedded-assemblies.cc
… src/monodroid/jni/embedded-assemblies.hh
… src/monodroid/jni/generate-pinvoke-tables.cc
… src/monodroid/jni/monodroid-glue-internal.hh
… src/monodroid/jni/monodroid-glue.cc
… src/monodroid/jni/pinvoke-tables.include