Use the hash to map any form of DSO name sent to us by Mono (we generate
hashes for several mutations) to find a cache entry for that DSO. Cache
includes the actual DSO name, whether it should be ignored when
requested (true for empty AOT DSOs) and the cached module handle.

This removes a bit of string processing from both startup and the
application run time.

This commit started with the goal to implement the above, but during 
implementation of the DSO cache I hit known limitations of the native
assembler generator code which, until this commit weren't a problem
(mostly related to hard-coded structure and array alignment).  So now
it includes a new implementation of the said generator, this time it 
fully implements the structure and symbol alignment calculation for
all the supported architectures.  Also added is the ability to generate
assembler code from managed structures, using reflection, without the need
of manually written code.  It also fixes previously unnoticed issue which
made typemap structures not aligned properly, which may have made them 
slightly slower than necessary

The implemented changes improve startup of both plain XA and MAUI applications,
measured on Pixel 3 XL running Android 12:

# Plain XA application

## AOT, Displayed time

| Before  | After   | Δ        | Notes                                 |
| ------- | ------- | -------- | ------------------------------------- |
| 252.400 | 245.000 | -2.93% ✓ | defaults; profiled AOT; 32-bit build  |
| 253.500 | 247.000 | -2.56% ✓ | defaults; profiled AOT; 64-bit build  |
| 251.000 | 247.100 | -1.55% ✓ | defaults; full AOT; 64-bit build      |
| 255.700 | 252.500 | -1.25% ✓ | defaults; full AOT; 32-bit build      |
| 257.000 | 256.400 | -0.23% ✓ | defaults; full AOT+LLVM; 64-bit build |
| 254.300 | 258.000 | +1.43% ✗ | defaults; full AOT+LLVM; 32-bit build |

## AOT, Total Initialization Time

| Before | After  | Δ        | Notes                                 |
| ------ | ------ | -------- | ------------------------------------- |
| 41.684 | 38.677 | -7.21% ✓ | defaults; profiled AOT; 64-bit build  |
| 41.357 | 38.727 | -6.36% ✓ | defaults; profiled AOT; 32-bit build  |
| 47.356 | 43.911 | -7.27% ✓ | defaults; full AOT; 32-bit build      |
| 47.256 | 44.356 | -6.14% ✓ | defaults; full AOT; 64-bit build      |
| 49.223 | 45.813 | -6.93% ✓ | defaults; full AOT+LLVM; 64-bit build |
| 48.755 | 46.098 | -5.45% ✓ | defaults; full AOT+LLVM; 32-bit build |

## No AOT, Displayed time

| Before  | After   | Δ        | Notes                                          |
| ------- | ------- | -------- | ---------------------------------------------- |
| 309.900 | 305.900 | -1.29% ✓ | preload disabled; no compression; 32-bit build |
| 317.900 | 309.900 | -2.52% ✓ | preload disabled; no compression; 64-bit build |
| 311.500 | 311.600 | +0.03% ✗ | preload enabled; 64-bit build                  |
| 314.800 | 311.600 | -1.02% ✓ | preload disabled; 64-bit build                 |
| 309.200 | 312.500 | +1.06% ✗ | preload enabled; no compression; 64-bit build  |
| 312.600 | 313.300 | +0.22% ✗ | preload disabled; 32-bit build                 |
| 314.100 | 315.300 | +0.38% ✗ | preload enabled; 32-bit build                  |
| 312.300 | 315.700 | +1.08% ✗ | preload enabled; no compression; 32-bit build  |

## No AOT, Total Init Time

| Before | After  | Δ        | Notes                                          |
| ------ | ------ | -------- | ---------------------------------------------- |
| 61.433 | 60.357 | -1.75% ✓ | preload disabled; no compression; 32-bit build |
| 60.245 | 60.521 | +0.46% ✗ | preload enabled; no compression; 32-bit build  |
| 60.216 | 60.801 | +0.96% ✗ | preload disabled; no compression; 64-bit build |
| 60.261 | 61.141 | +1.44% ✗ | preload enabled; no compression; 64-bit build  |
| 62.366 | 61.636 | -1.17% ✓ | preload enabled; 64-bit build                  |
| 61.924 | 61.776 | -0.24% ✓ | preload disabled; 64-bit build                 |
| 61.754 | 62.007 | +0.41% ✗ | preload enabled; 32-bit build                  |
| 62.529 | 62.792 | +0.42% ✗ | preload disabled; 32-bit build                 |

# MAUI Hello World

## AOT, Displayed time

| Before   | After    | Δ        | Notes                                 |
| -------- | -------- | -------- | ------------------------------------- |
| 868.900  | 831.700  | -4.28% ✓ | defaults; profiled AOT; 64-bit build  |
| 868.900  | 839.200  | -3.42% ✓ | defaults; profiled AOT; 32-bit build  |
| 901.900  | 868.800  | -3.67% ✓ | defaults; full AOT; 64-bit build      |
| 914.300  | 876.600  | -4.12% ✓ | defaults; full AOT+LLVM; 64-bit build |
| 904.300  | 877.500  | -2.96% ✓ | defaults; full AOT; 32-bit build      |
| 913.400  | 879.000  | -3.77% ✓ | defaults; full AOT+LLVM; 32-bit build |

## AOT, Total Init Time

| Before  | After   | Δ         | Notes                                 |
| ------- | ------- | --------- | ------------------------------------- |
| 114.953 | 95.573  | -16.86% ✓ | defaults; profiled AOT; 32-bit build  |
| 115.280 | 95.971  | -16.75% ✓ | defaults; profiled AOT; 64-bit build  |
| 205.505 | 172.190 | -16.21% ✓ | defaults; full AOT; 64-bit build      |
| 206.151 | 172.892 | -16.13% ✓ | defaults; full AOT; 32-bit build      |
| 215.617 | 181.895 | -15.64% ✓ | defaults; full AOT+LLVM; 64-bit build |
| 216.313 | 181.980 | -15.87% ✓ | defaults; full AOT+LLVM; 32-bit build |

## No AOT, Displayed time

| Before   | After    | Δ        | Notes                                          |
| -------- | -------- | -------- | ---------------------------------------------- |
| 1152.800 | 1145.200 | -0.66% ✓ | preload enabled; 64-bit build; no compression  |
| 1164.000 | 1153.100 | -0.94% ✓ | preload enabled; 64-bit build                  |
| 1141.500 | 1155.200 | +1.19% ✗ | preload disabled; 64-bit build; no compression |
| 1156.300 | 1159.500 | +0.28% ✗ | preload disabled; 64-bit build                 |
| 1230.500 | 1231.900 | +0.11% ✗ | preload enabled; 32-bit build; no compression  |
| 1235.600 | 1238.300 | +0.22% ✗ | preload disabled; 32-bit build; no compression |
| 1241.400 | 1241.900 | +0.04% ✗ | preload enabled; 32-bit build                  |
| 1240.500 | 1253.100 | +1.01% ✗ | preload disabled; 32-bit build                 |

## No AOT, Total Init Time

| Before | After  | Δ        | Notes                                          |
| ------ | ------ | -------- | ---------------------------------------------- |
| 73.570 | 73.827 | +0.35% ✗ | preload disabled; 64-bit build; no compression |
| 73.512 | 73.962 | +0.61% ✗ | preload enabled; 64-bit build; no compression  |
| 76.792 | 76.771 | -0.03% ✓ | preload enabled; 64-bit build                  |
| 77.382 | 77.109 | -0.35% ✓ | preload disabled; 64-bit build                 |
| 84.462 | 84.566 | +0.12% ✗ | preload disabled; 32-bit build; no compression |
| 84.051 | 84.725 | +0.80% ✗ | preload enabled; 32-bit build; no compression  |
| 88.051 | 88.015 | -0.04% ✓ | preload enabled; 32-bit build                  |
| 87.596 | 88.445 | +0.96% ✗ | preload disabled; 32-bit build                 |