@grendello 2022-01-14
Regarding the TODOs above, the differences in both mentioned cases are so small that I chalked it up to measurement error (especially for Displayed times which is inherently unreliable), but JIT may be affected by external factors on device as well. The tests are executed on a "live" device which does a lot in the background, any form of I/O may affect the tests performance. Over the time I've been working on performance, I learned that differences of around 1% can largely be ignored and accepted as "no change"